// Backend/Frontend/src/app/models/register.model.ts
export class register {
  email?: string;
  firstName?: string;
  lastName?: string;
  tel?: string;
  password?: string;
  confirmPassword?: string; // Add this line
  favoritePet?: string;
  bestFriend?: string;
}

export class passwordResponce {
  passwordResponceText?: string;
  passwordResponceImage?: string;
  passwordResponceKey?: string;
}